backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
apscheduler==3.10.4
pydantic==2.10.3
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
